refactor: separate auth context logic into dedicated utils file
