+++ apps/web/src/store/AuthContext.tsx
@@ -1,45 +1,10 @@
-import React, { createContext, useContext, useState, useEffect, ReactNode } from 'react';
+import React, { useState, useEffect, ReactNode } from 'react';
 import { User } from '../types';
 import { auth, clearJwtToken } from '../lib/api';
 import cookieService from '../lib/cookieService';
+import { AuthContext, mapApiRoleToClientRole } from './authUtils';
 
-/**
- * Authentication context interface
- */
-interface AuthContextType {
-  user: User | null;
-  requestVerificationCode: (email: string) => Promise<boolean>;
-  verifyCode: (email: string, code: string) => Promise<void>;
-  logout: () => Promise<void>;
-  isLoading: boolean;
-  error: string | null;
-  isAuthenticated: boolean;
-}
 
-/**
- * Create the auth context with default values
- */
-const AuthContext = createContext<AuthContextType>({
-  isAuthenticated: false,
-  isLoading: false,
-  user: null,
-  error: null,
-  requestVerificationCode: async () => false,
-  verifyCode: async () => {},
-  logout: async () => {},
-});
-
-export const useAuth = () => useContext(AuthContext);
-
-/**
- * Maps API role strings to client-side role enum values
- */
-function mapApiRoleToClientRole(apiRole: string): 'admin' | 'staff' | 'user' {
-  const role = apiRole.toUpperCase();
-  if (role === 'ADMIN') return 'admin';
-  if (role === 'STAFF') return 'staff';
-  return 'user';
-}
 
 /**
  * AuthProvider component that manages authentication state
